Rename joint and march_robot
