+++ march_hardware/CMakeLists.txt
@@ -54,6 +54,8 @@ add_library(${PROJECT_NAME}
     include/${PROJECT_NAME}/imotioncube/imotioncube.h
     include/${PROJECT_NAME}/imotioncube/imotioncube_state.h
     include/${PROJECT_NAME}/imotioncube/imotioncube_target_state.h
+    include/${PROJECT_NAME}/joint.h
+    include/${PROJECT_NAME}/march_robot.h
     include/${PROJECT_NAME}/power/boot_shutdown_offsets.h
     include/${PROJECT_NAME}/power/high_voltage.h
     include/${PROJECT_NAME}/power/low_voltage.h
@@ -62,8 +64,6 @@ add_library(${PROJECT_NAME}
     include/${PROJECT_NAME}/power/power_distribution_board.h
     include/${PROJECT_NAME}/temperature/temperature_ges.h
     include/${PROJECT_NAME}/temperature/temperature_sensor.h
-    src/Joint.cpp
-    src/MarchRobot.cpp
     src/encoder/absolute_encoder.cpp
     src/encoder/encoder.cpp
     src/encoder/incremental_encoder.cpp
@@ -75,6 +75,8 @@ add_library(${PROJECT_NAME}
     src/ethercat/sdo_interface.cpp
     src/imotioncube/imotioncube.cpp
     src/imotioncube/imotioncube_target_state.cpp
+    src/joint.cpp
+    src/march_robot.cpp
     src/power/high_voltage.cpp
     src/power/low_voltage.cpp
     src/power/power_distribution_board.cpp
