+++ march_hardware/src/joint.cpp
@@ -1,6 +1,6 @@
 // Copyright 2019 Project March.
 #include "march_hardware/ethercat/slave.h"
-#include "march_hardware/Joint.h"
+#include "march_hardware/joint.h"
 #include "march_hardware/error/hardware_exception.h"
 #include "march_hardware/error/motion_error.h"
